HV-1478 Move ValidationExtension to the exported package
