+++ cdi/src/main/java/org/hibernate/validator/internal/cdi/InheritedMethodsHelper.java
@@ -25,7 +25,7 @@ import org.hibernate.validator.internal.util.privilegedactions.GetMethods;
  * @author Gunnar Morling
  *
  */
-class InheritedMethodsHelper {
+public class InheritedMethodsHelper {
 
 	/**
 	 * Get a list of all methods which the given class declares, implements,
@@ -37,7 +37,7 @@ class InheritedMethodsHelper {
 	 *
 	 * @return set of all methods of the given class
 	 */
-	static List<Method> getAllMethods(Class<?> clazz) {
+	public static List<Method> getAllMethods(Class<?> clazz) {
 		Contracts.assertNotNull( clazz );
 
 		List<Method> methods = newArrayList();
