+++ cdi/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-org.hibernate.validator.internal.cdi.ValidationExtension
+org.hibernate.validator.cdi.ValidationExtension
